requirements.txt
torch==2.0.1
transformers==4.53.2
tensorflow==2.13.0
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.10.1
opencv-python==4.7.0.72
pytesseract==0.3.13
pymupdf==1.22.5
Pillow==9.5.0
typing-extensions==4.5.0
